+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,37 @@
 # gigadrive
 
+## 2.5.31
+
+### Patch Changes
+
+- Updated dependencies [[`ae7a4ab`](https://github.com/Gigadrive/sdk/commit/ae7a4abf09180d7e624d0931b24c3080e1e89a44)]:
+  - @gigadrive/network-config@4.3.4
+  - @gigadrive/sdk@0.7.25
+
+## 2.5.30
+
+### Patch Changes
+
+- Updated dependencies [[`50d11e6`](https://github.com/Gigadrive/sdk/commit/50d11e6d3912634aab1cef9c5e5923bfe6013de0)]:
+  - @gigadrive/network-config@4.3.3
+  - @gigadrive/sdk@0.7.24
+
+## 2.5.29
+
+### Patch Changes
+
+- Updated dependencies [[`e3c4946`](https://github.com/Gigadrive/sdk/commit/e3c494682ecedc8a7bbe0104648ced5c72e90b7e)]:
+  - @gigadrive/network-config@4.3.2
+  - @gigadrive/sdk@0.7.23
+
+## 2.5.28
+
+### Patch Changes
+
+- Updated dependencies [[`ddbd1f9`](https://github.com/Gigadrive/sdk/commit/ddbd1f9570856a962137e789dc59ab0eeb30f678)]:
+  - @gigadrive/network-config@4.3.1
+  - @gigadrive/sdk@0.7.22
+
 ## 2.5.27
 
 ### Patch Changes